add more runtime extensions
